% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grPipe.plot.R
\name{grPipe.plot}
\alias{grPipe.plot}
\title{Plot grPipe Nodes}
\usage{
grPipe.plot(
  nodes,
  pngfile,
  title = "",
  plot = TRUE,
  showGrid = FALSE,
  colSpace = 0.5,
  rowSpace = 0.5
)
}
\arguments{
\item{nodes}{data.frame}

\item{pngfile}{character}

\item{title}{character}

\item{plot}{logical}

\item{showGrid}{logical}

\item{colSpace}{numeric}

\item{rowSpace}{numeric}
}
\value{
No return value.
}
\description{
save grPipe nodes in \strong{pngfile} path.
}
\examples{
nodes = grPipe.create(2,5)
nodes = grPipe.node(nodes, "A1",  "A2",  "input")
nodes = grPipe.node(nodes, "A2",  "B2",  "step 1")
nodes = grPipe.node(nodes, "B2",  "B3",  "step 2")
nodes = grPipe.node(nodes, "B3",  "B4",  "step 3")
nodes = grPipe.node(nodes, "B4",  "A4",  "step 4")
nodes = grPipe.node(nodes, "A4",  "A5",  "step 5")
nodes = grPipe.node(nodes, "A5",  NA,  "output")
grPipe.plot(nodes, tempfile(), showGrid = TRUE)
grPipe.plot(nodes, tempfile(), showGrid = FALSE)

}
\author{
Daniel Gaspar Gonçalves
}
