% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grPipe.node.R
\name{grPipe.node}
\alias{grPipe.node}
\title{add or update grPipe nodes}
\usage{
grPipe.node(nodes, id, id_next, text)
}
\arguments{
\item{nodes}{data.frame}

\item{id}{character}

\item{id_next}{character}

\item{text}{character}
}
\description{
parameters:
    - nodes: data.frame (
          id: character,
          id_next: character,
          text: character
      )
    - id: character
    - id_next: character
    - text: character
return data.frame colnames:
    - id: character
    - id_next: character
    - text: character
add a new node if it doesn't exist or update an existing one.
}
\examples{
nodes = grPipe.create(2,5)
nodes = grPipe.node(nodes, "A1",  "A2",  "input")
nodes = grPipe.node(nodes, "A2",  "B2",  "step 1")
nodes = grPipe.node(nodes, "B2",  "B3",  "step 2")
nodes = grPipe.node(nodes, "B3",  "B4",  "step 3")
nodes = grPipe.node(nodes, "B4",  "A4",  "step 4")
nodes = grPipe.node(nodes, "A4",  "A5",  "step 5")
nodes = grPipe.node(nodes, "A5",  NA,  "output")

}
\author{
Daniel Gaspar Gonçalves
}
